Running Linux with new SS
git-svn-id: svn://136.177.114.72/svn_GW/phreeqcpp/branches/ErrorHandling/src@6075 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ phreeqc/prep.cpp
@@ -799,7 +799,7 @@ build_ss_assemblage(void)
 			row = x[i]->number * (count_unknowns + 1);
 			for (size_t j = 0; j < ss_ptr->Get_ss_comps().size(); j++)
 			{
-				if (j != x[i]->ss_comp_number)
+				if ((int) j != x[i]->ss_comp_number)
 				{
 /*					store_jacob (&(s_s_ptr->dn), &(array[row + col + j]), -1.0); */
 					store_jacob(&(x[i]->phase->dn), &(array[row + col + j]),
